Change the way parameters are passed to subscriber's functions
Now the src_property_name parameter cannot be 'None'. There cannot be
registration to the event hub made with such property name so
passing 'None' was pointless.
Now the function tries at first to get the getter method from
source property and if it cannot be found, it tries to get
the method directly from source base object.